refactor: 도메인 중심으로 백엔드 디렉토리 구조 정리
- app/api 모듈을 app/domains 하위로 이동
- app/core/security.py를 app/common/security.py로 이동
- 사용하지 않는 app/legacy 디렉토리 삭제
- main.py와 테스트의 import 경로 수정
- 기존 테스트 통과 확인
