+++ app/common/dependency/dependencies.py
@@ -7,7 +7,7 @@ from typing import Optional
 from app.infrastructure.external.openai_client import OpenAIClient
 from app.infrastructure.db.vector.vector_db import get_vector_db
 from app.infrastructure.embedding.embedding_model import get_embedding_model
-from app.api.chat.service.chat_service import ChatService
+from app.domains.chat.service import ChatService
 from app.common.logging.logging_config import get_logger
 
 logger = get_logger(__name__)
