--- app/core/README.md
@@ -1,11 +1,0 @@
-# Core
-
-애플리케이션 전역 관심사를 둔다.
-
-예정 역할:
-- 설정 로딩과 검증
-- 인증/토큰 보안 유틸
-- 공통 dependency
-- 로깅 설정
-
-기존 `app/common` 코드는 당장 옮기지 않는다. 새 기능부터 `app/core`를 사용하고, 이후 안정화된 것만 점진적으로 이동한다.
